Move currency update logic into it's own concern.
