+++ core/app/models/spree/order.rb
@@ -4,6 +4,7 @@ require 'spree/order/checkout'
 module Spree
   class Order < ActiveRecord::Base
     include Checkout
+    include CurrencyUpdater
 
     checkout_flow do
       go_to_state :address
@@ -613,37 +614,6 @@ module Spree
       def set_currency
         self.currency = Spree::Config[:currency] if self[:currency].nil?
       end
-  
-      # Updates the order's prices and all line items prices
-      def homogenize_line_item_currencies
-        begin
-          update! if update_line_item_currencies!
-        rescue
-          return false # flunk the validation
-        end
-      end
-                                                    
-      # Updates prices of order's line items
-      def update_line_item_currencies!
-        line_items.where('currency != ?', currency).each do |line_item|
-          update_line_item_price!(line_item)
-        end
-      end
-  
-      # Returns the price object from given item
-      def price_from_line_item(line_item)
-        line_item.variant.prices.where(currency: currency).first
-      end
-  
-      # Updates price from given line item
-      def update_line_item_price!(line_item)
-        price = price_from_line_item(line_item)
 
-        if price
-          line_item.update_attributes!(currency: price.currency, price: price.amount)
-        else
-          raise 'no price found for line item'
-        end
-      end
   end
 end
